Updating ETL scripts
Simplifying scripts to allow just one script to handle all taxi data from web to GCS and one script to handle all data from GCS to BQ
