+++ Prefect/Old Scripts/docker_deploy.py
@@ -1,0 +1,16 @@
+
+from prefect.deployments import Deployment
+from parameterized_flow import etl_parent_flow
+from prefect.infrastructure.docker import DockerContainer
+
+docker_block = DockerContainer.load("zoom")
+
+docker_dep = Deployment.build_from_flow(
+    flow=etl_parent_flow,
+    name="docker-flow",
+    infrastructure=docker_block,
+)
+
+
+if __name__ == "__main__":
+    docker_dep.apply()
