+++ prefect/parameterized_flow.py
@@ -6,27 +6,87 @@ from prefect.tasks import task_input_hash
 from datetime import timedelta
 
 @task(retries=3)
-def fetch(dataset_url: str) -> pd.DataFrame:
-    """Read data from web into pandas DataFrame"""
-
-    df = pd.read_csv(dataset_url)
-    return df
-
+def fetch(dataset_url: str, service: str) -> pd.DataFrame:
+    """Read data from web into pandas DataFrame. Enforce data types."""
+    if service == "yellow":
+        df = pd.read_csv(dataset_url).astype({
+            "VendorID": "Int64",
+            "tpep_pickup_datetime": "object",
+            "tpep_dropoff_datetime": "object",
+            "store_and_fwd_flag": "object",
+            "RatecodeID": "Int64",
+            "PULocationID": "Int64",
+            "DOLocationID": "Int64",
+            "payment_type": "Int64",
+            "passenger_count": "Int64",
+            "trip_distance": "Float64",
+            "fare_amount": "Float64",
+            "extra": "Float64",
+            "mta_tax": "Float64",
+            "tip_amount": "Float64",
+            "tolls_amount": "Float64",
+            "improvement_surcharge": "Float64",
+            "total_amount": "Float64",
+            "congestion_surcharge": "Float64" 
+         })
+        return df
+    elif service == "green":
+        df = pd.read_csv(dataset_url).astype({
+            "VendorID": "Int64",
+            "lpep_pickup_datetime": "object",
+            "lpep_dropoff_datetime": "object",
+            "store_and_fwd_flag": "object",
+            "RatecodeID": "Int64",
+            "PULocationID": "Int64",
+            "DOLocationID": "Int64",
+            "passenger_count": "Int64",
+            "trip_distance": "Float64",
+            "fare_amount": "Float64",
+            "extra": "Float64",
+            "mta_tax": "Float64",
+            "tip_amount": "Float64",
+            "tolls_amount": "Float64",
+            "ehail_fee": "Float64",
+            "improvement_surcharge": "Float64",
+            "total_amount": "Float64",
+            "payment_type": "Int64",
+            "trip_type": "Int64",
+            "congestion_surcharge": "Float64" 
+         })
+        return df
+    elif service == "fhv":
+        df = pd.read_csv(dataset_url).astype({
+            "dispatching_base_num": "object",
+            "pickup_datetime": "object",
+            "dropOff_datetime": "object",
+            "PUlocationID": "Int64",
+            "DOlocationID": "Int64",
+            "SR_Flag": "Int64",
+            "Affiliated_base_number": "object"
+        })
+        return df
 
 @task(log_prints=True)
-def clean(df=pd.DataFrame) -> pd.DataFrame:
+def clean(df: pd.DataFrame, service: str) -> pd.DataFrame:
     """Fix dtype issues"""
-    df['tpep_pickup_datetime'] = pd.to_datetime(df['tpep_pickup_datetime'])
-    df['tpep_dropoff_datetime'] = pd.to_datetime(df['tpep_dropoff_datetime'])
-    print(df.head(2))
-    print(f"columns: {df.dtypes}")
-    print(f"rows: {len(df)}")
-    return df
+    if service == "yellow":
+        df['tpep_pickup_datetime'] = pd.to_datetime(df['tpep_pickup_datetime'])
+        df['tpep_dropoff_datetime'] = pd.to_datetime(df['tpep_dropoff_datetime'])
+        return df
+    elif service == "green":
+        df['lpep_pickup_datetime'] = pd.to_datetime(df['lpep_pickup_datetime'])
+        df['lpep_dropoff_datetime'] = pd.to_datetime(df['lpep_dropoff_datetime'])
+        return df  
+    elif service == "fhv":
+        df['pickup_datetime'] = pd.to_datetime(df['pickup_datetime'])
+        df['dropOff_datetime'] = pd.to_datetime(df['dropOff_datetime'])
+        return df
+
 
 @task()
-def write_local(df: pd.DataFrame, color: str, dataset_file:str) -> Path:
+def write_local(df: pd.DataFrame, service: str, dataset_file:str) -> Path:
     """Write dataframe as parquet file """
-    path = Path(f"data/{color}/{dataset_file}.parquet").as_posix()
+    path = Path(f"data/{service}/{dataset_file}.parquet").as_posix()
     df.to_parquet(path, compression = 'gzip')
     return path
 
@@ -41,27 +101,27 @@ def write_gcs(path: Path) -> None:
     return
 
 @flow()
-def etl_web_to_gcs(year: int, month: int, color: str) -> None:
+def etl_web_to_gcs(year: int, month: int, service: str) -> None:
     """The main ETL function"""
-    dataset_file = f"{color}_tripdata_{year}-{month:02}"
-    dataset_url = f"https://github.com/DataTalksClub/nyc-tlc-data/releases/download/{color}/{dataset_file}.csv.gz"
+    dataset_file = f"{service}_tripdata_{year}-{month:02}"
+    dataset_url = f"https://github.com/DataTalksClub/nyc-tlc-data/releases/download/{service}/{dataset_file}.csv.gz"
 
-    df = fetch(dataset_url)
-    df_clean = clean(df)
-    path = write_local(df_clean, color, dataset_file)
+    df = fetch(dataset_url, service)
+    df_clean = clean(df, service)
+    path = write_local(df_clean, service, dataset_file)
     write_gcs(path)
 
 
 @flow()
 def etl_parent_flow(
-    months: list[int] = [1,2], year: int = 2021, color: str = "yellow"
+    months: list[int] = [1,2], year: int = 2021, service: str = "yellow"
 ):
     for month in months:
-        etl_web_to_gcs(year, month, color)
+        etl_web_to_gcs(year, month, service)
 
 if __name__ == '__main__':
-    color = "yellow"
+    service = "yellow"
     months = [1,2,3]
     year = 2019
-    etl_parent_flow(months,year,color)
+    etl_parent_flow(months,year,service)
